Added embedding metrics and changed name
